requirements.txt
pandas>=2.2
matplotlib>=3.8
python-dotenv>=1.0
openai>=1.30
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
duckdb>=0.10
sqlalchemy>=2.0
jupyter>=1.0
tabulate>=0.9
plotly>=5.22
rich>=13.7
